Fix preset variable

+++ wled00/wled.h
@@ -564,7 +564,7 @@ WLED_GLOBAL JsonDocument* fileDoc;
 WLED_GLOBAL bool doCloseFile _INIT(false);
 
 // presets
-WLED_GLOBAL int8_t currentPreset _INIT(-1);
+WLED_GLOBAL int16_t currentPreset _INIT(-1);
 
 WLED_GLOBAL byte errorFlag _INIT(0);
